fix: WIP on addressing v_end that is outside write region in mark_keep_variants and reconstruct alg.

+++ tests/dataset/genotypes/test_reconstruct.py
@@ -1,6 +1,6 @@
 # %%
 import numpy as np
-from genvarloader.dataset.genotypes import (
+from genvarloader._dataset._genotypes import (
     SparseGenotypes,
     reconstruct_haplotype_from_sparse,
 )
@@ -15,7 +15,7 @@ def case_snps():
     # (s p v)
     genos = np.array([[[0, 1]]], dtype=np.int8)
     first_v_idxs = np.array([0], dtype=np.int32)
-    offsets = np.array([0, 2], np.int32)
+    offsets = np.array([0, 2], np.int64)
 
     shift = 0
     alt_alleles = np.frombuffer(b"AT", dtype=np.uint8)
@@ -48,7 +48,7 @@ def case_indels():
     # (s p v) : (1 1 2)
     genos = np.array([[[1, 1]]], dtype=np.int8)
     first_v_idxs = np.array([0], dtype=np.int32)
-    offsets = np.array([0, 2], np.int32)
+    offsets = np.array([0, 2], np.int64)
 
     shift = 0
     alt_alleles = np.frombuffer(b"GAT", dtype=np.uint8)
@@ -81,7 +81,7 @@ def case_spanning_del():
     # (s p v) : (1 1 2)
     genos = np.array([[[1]]], dtype=np.int8)
     first_v_idxs = np.array([0], dtype=np.int32)
-    offsets = np.array([0, 1], np.int32)
+    offsets = np.array([0, 1], np.int64)
 
     shift = 0
     alt_alleles = np.frombuffer(b"G", dtype=np.uint8)
@@ -114,7 +114,7 @@ def case_shift_ins():
     # (s p v) : (1 1 2)
     genos = np.array([[[1, 1]]], dtype=np.int8)
     first_v_idxs = np.array([0], dtype=np.int32)
-    offsets = np.array([0, 2], np.int32)
+    offsets = np.array([0, 2], np.int64)
 
     shift = 1
     alt_alleles = np.frombuffer(b"TCGA", dtype=np.uint8)
